pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dataagentbench"
version = "0.1.0"
description = "A benchmark for evaluating LLM agents on data science tasks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.25.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "RestrictedPython>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
dab = "dataagentbench.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["dataagentbench"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

[tool.coverage.run]
source = ["dataagentbench"]
omit = ["tests/*"]
